+++ src/core/energy_storage/electric_storage.jl
@@ -167,7 +167,7 @@ end
     model_degradation::Bool = false
     degradation::Dict = Dict()
     minimum_avg_soc_fraction::Float64 = 0.0
-    per_timestep_self_discharge_fraction::Float64 = 0.0 # Battery self-discharge as a fraction per timestep loss based on kWh stored in each timestep
+    self_discharge_fraction_per_timestep::Float64 = 0.0 # Battery self-discharge as a fraction per timestep loss based on kWh stored in each timestep
     is_ldes::Bool = false
     duration::Real = 100.0
     require_start_and_end_charge_to_be_equivalent::Bool = false
@@ -206,7 +206,7 @@ Base.@kwdef struct ElectricStorageDefaults
     model_degradation::Bool = false
     degradation::Dict = Dict()
     minimum_avg_soc_fraction::Float64 = 0.0
-    per_timestep_self_discharge_fraction::Float64 = 0.0
+    self_discharge_fraction_per_timestep::Float64 = 0.0
     is_ldes::Bool = false
     duration::Real = 100.0
     require_start_and_end_charge_to_be_equivalent::Bool = false
@@ -253,7 +253,7 @@ struct ElectricStorage <: AbstractElectricStorage
     model_degradation::Bool
     degradation::Degradation
     minimum_avg_soc_fraction::Float64
-    per_timestep_self_discharge_fraction::Float64
+    self_discharge_fraction_per_timestep::Float64
     is_ldes::Bool
     duration::Real
     require_start_and_end_charge_to_be_equivalent::Bool
@@ -347,7 +347,7 @@ struct ElectricStorage <: AbstractElectricStorage
             s.model_degradation,
             degr,
             s.minimum_avg_soc_fraction,
-            s.per_timestep_self_discharge_fraction,
+            s.self_discharge_fraction_per_timestep,
             s.is_ldes,
             s.duration,
             s.require_start_and_end_charge_to_be_equivalent